requirements.txt
aiogram==3.1.1
python-dotenv
gspread==6.1.2
redis==5.0.7
pydantic==2.3
pandas